classification-service/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
transformers==5.7.0
torch==2.11.0+cpu
pillow==12.2.0
python-multipart==0.0.27

--extra-index-url https://download.pytorch.org/whl/cpu
